ppt-generator-ui/backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
pandas
PyPDF2
python-docx
groq
openpyxl
